refactor: modify code style
