+++ src/main/java/roomescape/controller/AdminThemeController.java
@@ -17,7 +17,7 @@ import roomescape.service.AdminThemeService;
 public class AdminThemeController {
     private final AdminThemeService adminThemeService;
 
-    public AdminThemeController(AdminThemeService adminThemeService){
+    public AdminThemeController(AdminThemeService adminThemeService) {
         this.adminThemeService = adminThemeService;
     }
