--- src/test/java/roomescape/RoomescapeApplicationTest.java
@@ -1,13 +1,0 @@
-package roomescape;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RoomescapeApplicationTest {
-
-        @Test
-        void contextLoads() {
-        }
-
-}
