+++ src/test/java/roomescape/controller/AdminReservationControllerTest.java
@@ -4,7 +4,6 @@ import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.annotation.DirtiesContext;
@@ -13,10 +12,6 @@ import org.springframework.test.annotation.DirtiesContext.ClassMode;
 @SpringBootTest(webEnvironment = WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
 class AdminReservationControllerTest {
-
-    @Autowired
-    private AdminReservationController controller;
-
     @DisplayName("모든 사용자의 예약 내역이 모두 조회되어야한다.")
     @Test
     void 관리자_예약_조회_API() {
@@ -26,8 +21,4 @@ class AdminReservationControllerTest {
                 .then().log().all()
                 .statusCode(200);
     }
-
-
-
-
 }
