+++ src/test/java/roomescape/controller/AdminThemeControllerTest.java
@@ -1,14 +1,11 @@
 package roomescape.controller;
 
-import static org.hamcrest.Matchers.equalTo;
-
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.annotation.DirtiesContext;
@@ -17,11 +14,10 @@ import org.springframework.test.annotation.DirtiesContext.ClassMode;
 @SpringBootTest(webEnvironment = WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
 class AdminThemeControllerTest {
-
     @DisplayName("관리자 테마 추가")
     @Test
-    void 관리자_테마_추가(){
-        Map<String,Object> params = new HashMap<>();
+    void 관리자_테마_추가() {
+        Map<String, Object> params = new HashMap<>();
         params.put("name", "공포의 폐병원");
         params.put("description", "공포의 폐병원");
         params.put("thumbnail_url", "https://images.unsplash.com/photo-1505635552518-3448ff116af3?w=300&q=80");
@@ -37,7 +33,7 @@ class AdminThemeControllerTest {
     @DisplayName("API - 관리자 테마 삭제")
     @Test
     void API_관리자_테마_삭제() {
-        Map<String,Object> params = new HashMap<>();
+        Map<String, Object> params = new HashMap<>();
         params.put("name", "공포의 폐병원");
         params.put("description", "공포의 폐병원");
         params.put("thumbnail_url", "https://images.unsplash.com/photo-1505635552518-3448ff116af3?w=300&q=80");
