+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -6,7 +6,6 @@ import java.util.HashMap;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.annotation.DirtiesContext;
@@ -15,13 +14,10 @@ import org.springframework.test.annotation.DirtiesContext.ClassMode;
 @SpringBootTest(webEnvironment = WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
 class ReservationControllerTest {
-    @Autowired
-    private ReservationController reservationController;
-
     @DisplayName("사용자 예약 추가")
     @Test
     void 사용자_예약_추가_API() {
-        Map<String,Object> params = new HashMap<>();
+        Map<String, Object> params = new HashMap<>();
         params.put("name", "브라운");
         params.put("date", "2026-05-05");
         params.put("timeId", 1);
@@ -37,10 +33,10 @@ class ReservationControllerTest {
 
     @DisplayName("사용자 예약 삭제")
     @Test
-    void 사용자_예약_삭제(){
+    void 사용자_예약_삭제() {
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
-                .pathParam("id",1)
+                .pathParam("id", 1)
                 .when().delete("/reservations/{id}")
                 .then().log().all()
                 .statusCode(204);
@@ -48,15 +44,12 @@ class ReservationControllerTest {
 
     @DisplayName("사용자 예약 조회")
     @Test
-    void 사용자_예약_조회(){
+    void 사용자_예약_조회() {
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
-                .queryParam("username","김철수")
+                .queryParam("username", "김철수")
                 .when().get("/reservations")
                 .then().log().all()
                 .statusCode(200);
-
-
     }
-
-    }
+}
