+++ src/test/java/roomescape/controller/ReservationTimeControllerTest.java
@@ -16,21 +16,20 @@ import org.springframework.test.annotation.DirtiesContext.ClassMode;
 @SpringBootTest(webEnvironment = WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
 class ReservationTimeControllerTest {
-
     @DisplayName("API - 예약 시간 등록")
     @Test
     void API_예약_시간_등록() {
         final String createStartAt = "23:00";
-        final Map<String,Object> params = new HashMap<>();
+        final Map<String, Object> params = new HashMap<>();
         params.put("startAt", createStartAt);
 
         RestAssured.given().log().all()
-               .contentType(ContentType.JSON)
-               .body(params)
-               .when().post("/times")
-               .then().log().all()
-               .statusCode(201)
-               .body("startAt", equalTo(createStartAt));
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/times")
+                .then().log().all()
+                .statusCode(201)
+                .body("startAt", equalTo(createStartAt));
     }
 
     @DisplayName("API - 예약 시간 조회")
@@ -48,7 +47,7 @@ class ReservationTimeControllerTest {
     @Test
     void API_예약_시간_삭제() {
         final String createStartAt = "23:00";
-        final Map<String,Object> params = new HashMap<>();
+        final Map<String, Object> params = new HashMap<>();
         params.put("startAt", createStartAt);
 
         final long id = RestAssured.given().log().all()
