+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -6,9 +6,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.Theme;
 
 class ReservationTest {
 
@@ -20,20 +17,6 @@ class ReservationTest {
         this.theme = new Theme(1L, "공포의 저택", "버려진 저택에서 탈출하라! 어둠 속에 숨겨진 비밀을 밝혀야 살 수 있다.",
                 "https://picsum.photos/seed/haunted/400/250");
     }
-
-    //    private final Long id;
-    //    private final String name;
-    //    private final LocalDate date;
-    //    private final ReservationTime time;
-    //    private final Theme theme;
-
-    //public class Theme {
-    //    private final Long id;
-    //    private final String name;
-    //    private final String description;
-    //    private final String thumbnailUrl;
-
-    // '공포의 저택', '버려진 저택에서 탈출하라! 어둠 속에 숨겨진 비밀을 밝혀야 살 수 있다.', 'https://picsum.photos/seed/haunted/400/250');
 
     @Test
     void 예약_생성() {
