+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -1,10 +1,9 @@
-package roomescape.reservationtime;
+package roomescape.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalTime;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.ReservationTime;
 
 class ReservationTimeTest {
